+++ htdocs/components/KPIReport/KPIReportComponent.js
@@ -18,11 +18,15 @@ import DailySourceComponent from './subComponents/DailySourceComponent';
 
 import {groupBy} from '../utils/utils';
 
+import BillingConsultancyComponent from './subComponents/BillingConsultancyComponent';
+import APIUser from '../services/APIUser';
+import Toggle from '../shared/Toggle';
 
 export const ReportType = {Daily: "day", Weekly: "week", Monthly: "month"}
 
 export default class KPIReportComponent extends MainComponent {
     api;
+    apiUsers=new APIUser();
     ResultType;
     colors;
     reports;
@@ -34,6 +38,7 @@ export default class KPIReportComponent extends MainComponent {
     REP_SERVICE_REQUEST = 5;
     REP_SERVICE_REQUEST_SOURCE = 6;
     apiSDManagerDashboard = new APISDManagerDashboard();
+    REP_CONFIRMED_BILLED_PER_ENGINEER=5;
 
     /**
      * SRS_BY_CONTRAC
@@ -51,18 +56,28 @@ export default class KPIReportComponent extends MainComponent {
                 from: this.getInitStartDate(),
                 to: this.getInitEndDate(),
                 resultType: this.ResultType.Weekly,
-                customerID: ''
+                customerID: '',
+                consName:'',
+                teams:{
+                    hd:true,
+                    es:true,
+                    sp:true,
+                    p:true
+                }
             },
             data: [],
             reports: [],
-            activeReport: null
+            activeReport: null,
+            consultants:[]
         };
         this.api = new APIKPIReport();
         this.reportparameters = {
             dateFrom: 'dateFrom',
             dateTo: 'dateTo',
             customer: 'customer',
-            resultType: 'resultType'
+            resultType: 'resultType',
+            consName:'consName',
+            teams:'teams'
         };
         moment.locale("en");
         moment.updateLocale("en", {
@@ -74,6 +89,7 @@ export default class KPIReportComponent extends MainComponent {
 
     componentDidMount() {
         this.getReports();
+
     }
 
     getReports = () => {
@@ -139,11 +155,20 @@ export default class KPIReportComponent extends MainComponent {
                     this.reportparameters.resultType,
                 ]
             },
+            {
+                id: this.REP_CONFIRMED_BILLED_PER_ENGINEER,
+                title: "Billed Consultancy By Person",
+                parameters: [
+                    this.reportparameters.dateFrom,
+                    this.reportparameters.dateTo,
+                    this.reportparameters.consName,
+                    this.reportparameters.teams
+                ]
+            }
         ];
         reports.sort((a, b) => a.title.localeCompare(b.title));
         if (!activeReport)
-            activeReport = reports[0];
-
+            activeReport = reports[4];
         this.setState({reports, activeReport}, () => this.handleReportView());
     }
 
@@ -162,8 +187,7 @@ export default class KPIReportComponent extends MainComponent {
     };
 
     getFilterElement = () => {
-        const {filter, reports} = this.state;
-
+        const {filter, reports,activeReport,consultants} = this.state;
         return (
             <table>
                 <tbody>
@@ -172,6 +196,7 @@ export default class KPIReportComponent extends MainComponent {
                     <td colSpan="5">
                         <select style={{width: 180}}
                                 onChange={this.handleReportChange}
+                                value={activeReport?.id}
                         >
                             {reports.map((r) => (
                                 <option key={r.id}
@@ -238,6 +263,54 @@ export default class KPIReportComponent extends MainComponent {
                     ) : null}
                 </tr>
                 <tr>
+                    {this.hasParameter(this.reportparameters.teams) ? (
+                        <React.Fragment>
+                            <td>Teams</td>
+                            <td colSpan={3}>
+                               <div>
+
+                            <label className="mr-3 ml-5">HD</label>
+                            <Toggle checked={filter.teams.hd}
+                                    onChange={(value) => this.setFilterTeam("hd", !filter.teams.hd)}
+                            />
+                            <label className="mr-3 ml-5">ES</label>
+                            <Toggle checked={filter.teams.es}
+                                    onChange={(value) => this.setFilterTeam("es", !filter.teams.es)}
+                            />
+                            <label className="mr-3 ml-5">SP</label>
+                            <Toggle checked={filter.teams.sp}
+                                    onChange={(value) => this.setFilterTeam("sp", !filter.teams.sp)}
+                            />
+                            <label className="mr-3 ml-5">P</label>
+                            <Toggle checked={filter.teams.p}
+                                    onChange={(value) => this.setFilterTeam("p", !filter.teams.p)}
+                            />
+
+                               </div>
+                            </td>
+                        </React.Fragment>
+                    ) : null}
+                    <td>
+                     </td>
+                </tr>
+                <tr>
+                    {this.hasParameter(this.reportparameters.consName) ? (
+                        <React.Fragment>
+                            <td>Consultant</td>
+                            <td colSpan={3}>
+                               <select style={{width:340}} value={filter.consName}
+                               onChange={(event)=>this.setFilter('consName',event.target.value)}
+                               >
+                                <option>All</option>
+                                {consultants.map(c=><option key={c.id} value={c.name}>{c.name}</option>)}
+                               </select>
+                            </td>
+                        </React.Fragment>
+                    ) : null}
+                    <td>
+                     </td>
+                </tr>
+                <tr>
                     {this.hasParameter(this.reportparameters.customer) ? (
                         <React.Fragment>
                             <td>Customer</td>
@@ -261,6 +334,11 @@ export default class KPIReportComponent extends MainComponent {
             </table>
         );
     };
+    setFilterTeam=(team,value)=>{
+        const {filter}=this.state;
+        filter.teams[team]=value;
+        this.setState({filter});
+    }
     handleReportChange = ($event) => {
         const id = $event.target.value;
         const {reports} = this.state;
@@ -271,14 +349,20 @@ export default class KPIReportComponent extends MainComponent {
         const {activeReport} = this.state;
         return activeReport != null && activeReport.parameters.indexOf(parameter) >= 0;
     }
-    handleReportView = () => {
-        const {filter, activeReport} = this.state;
+    handleReportView = async() => {
+        let {filter, activeReport,consultants} = this.state;
+
         this.setState({_showSpinner: true});
         // if (filter.from == "") {
         //     this.alert("You must enter the start date");
         //     return;
-        // }        
-
+        // }
+if(activeReport&&activeReport.parameters.indexOf(this.reportparameters.consName)>=0&&consultants.length==0)
+        {
+            //console.log('have users');
+            consultants=await this.apiUsers.getActiveUsers();
+            this.setState({consultants});
+        }
         switch (activeReport?.id) {
             case this.REP_SR_FIXED:
                 this.api.getSRFixed(filter.from, filter.to, filter.customerID).then((data) => {
@@ -337,11 +421,17 @@ export default class KPIReportComponent extends MainComponent {
 
                 });
                 break;
+            case this.REP_CONFIRMED_BILLED_PER_ENGINEER:
+                this.api.getEngineerMonthlyBilling(filter.from, filter.to).then((data) => {
+                    this.processData(data, false,false,false);
+                });
+                break;
 
         }
 
+
     };
-    processData = (data, skipWeekEnds = true) => {
+    processData = (data, skipWeekEnds = true,daily=true,weekly=true,monthly=true) => {
         const {filter} = this.state;
         if (skipWeekEnds) {
             data = data.filter((a) => {
@@ -350,14 +440,15 @@ export default class KPIReportComponent extends MainComponent {
                 else return true;
             });
         }
-        if (data.length <= 100) filter.resultType = this.ResultType.Daily;
-        if (data.length > 100 && data.length < 200)
+        if (data.length <= 100&&daily) filter.resultType = this.ResultType.Daily;
+        if (data.length > 100 && data.length < 200&&weekly)
             filter.resultType = this.ResultType.Weekly;
-        else filter.resultType = this.ResultType.Monthly;
+        else if(monthly)
+        filter.resultType = this.ResultType.Monthly;
         this.setState({data, _showSpinner: false, filter});
     }
     getActiveChart = () => {
-        const {activeReport, data, filter} = this.state;
+        const {activeReport, data, filter,consultants} = this.state;
         switch (activeReport?.id) {
             case this.REP_SR_FIXED:
                 return <SRFixedComponent data={data}
@@ -384,6 +475,10 @@ export default class KPIReportComponent extends MainComponent {
                 return <DailySourceComponent data={data}
                                              filter={filter}
                 />
+            case this.REP_CONFIRMED_BILLED_PER_ENGINEER:
+                return <BillingConsultancyComponent data={data}
+                                                    filter={filter} consultants={consultants}>
+                </BillingConsultancyComponent>;
             default:
                 return null;
         }