+++ htdocs/components/KPIReport/services/APIKPIReport.js
@@ -25,4 +25,14 @@ export default class APIKPIReport extends APIMain {
     getDailySource(from,to,customerID){
         return this.get(`${ApiUrls.KPIReport}dailySource&from=${from}&to=${to}&customerID=${customerID}`);
     }
+    getEngineerMonthlyBilling(startDate,endDate){
+        return this.get(`${ApiUrls.KPIReport}engineerMonthlyBilling&from=${startDate}&to=${endDate}`).then(data=>{
+            data.map(d=>{
+                d.date=d.inh_date_printed_yearmonth.substring(0,4)+'-'+d.inh_date_printed_yearmonth.substring(4,6)+'-01';
+                d.engineer=d.inl_desc.replace(" - Consultancy","");
+                return d;
+            });
+            return data;
+        });
+    }
 }